# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/user-media/user/{userId}": {
      "get": {
        "operationId": "UserMediaController_getMediaByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserMediaResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserMedia"
        ]
      }
    },
...
    "/user-tributes/user/{userId}": {
      "get": {
        "operationId": "UserTributeController_getTributesByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserTributeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserTribute"
        ]
      }
    },
...
    "/user-featured-memorials/user/{userId}": {
      "get": {
        "operationId": "UserFeaturedMemorialController_getFeaturedMemorialsByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadUserFeaturedMemorialResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "summary": "Get all featured memorials for a user",
        "tags": [
          "UserFeaturedMemorial"
        ]
      }
    },
...
    "/userNotificationSettings/user/{userId}": {
      "get": {
        "operationId": "UserNotificationSettingsController_getByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserNotificationSettingResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserNotificationSettings"
        ]
      }
    },
...
    "/user-answers": {
      "post": {
        "operationId": "UserAnswersController_createAnswer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserAnswerRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserAnswersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserAnswers"
        ]
      },
      "get": {
        "operationId": "UserAnswersController_getAllAnswers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserAnswerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserAnswers"
        ]
      }
    },
    "/user-answers/{id}": {
      "patch": {
        "operationId": "UserAnswersController_updateAnswer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserAnswerRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserAnswersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserAnswers"
        ]
      },
      "delete": {
        "operationId": "UserAnswersController_deleteAnswer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserAnswersResponseDto"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserAnswers"
        ]
      }
    },
    "/user-answers/readById/{id}": {
      "get": {
        "operationId": "UserAnswersController_getAnswerById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserAnswerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserAnswers"
        ]
      }
    },
...
  },
  "info": {
    "title": "rememeber-me",
    "description": "API documentation for rememeber-me.env",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserRequestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "string"
          },
          "callingCode": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "photoId": {
            "type": "string"
          },
          "photoURL": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "street1": {
            "type": "string"
          },
          "street2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          }
        }
      },
      "UserRoles": {
        "type": "string",
        "enum": [
          "admin",
          "superAdmin",
          "user"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "callingCode": {
            "type": "string"
          },
          "defaultRole": {
            "type": "string"
          },
          "street1": {
            "type": "string"
          },
          "street2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "photoId": {
            "type": "string"
          },
          "photoURL": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRoles"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadLandingModeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-landing-mode"
          },
          "title": {
            "type": "string",
            "example": "Dark Mode"
          },
          "description": {
            "type": "string",
            "example": "Landing mode for dark theme"
          },
          "landingModeType": {
            "type": "string",
            "example": "full-mode",
            "enum": [
              "full-mode",
              "event-mode",
              "video-only-mode"
            ]
          },
          "iconId": {
            "type": "string",
            "example": "icon-file-id"
          },
          "iconURL": {
            "type": "string",
            "example": "https://cdn.example.com/icons/darkmode.png"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "title",
          "landingModeType",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadUserMediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-media"
          },
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "fileURL": {
            "type": "string",
            "example": "https://cdn.example.com/uploads/media-file.png"
          },
          "fileId": {
            "type": "string",
            "example": "uuid-of-file"
          },
          "type": {
            "type": "string",
            "example": "IMAGE",
            "description": "Media type (e.g., IMAGE, VIDEO, AUDIO)"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "photoCaption": {
            "type": "string",
            "example": "Vacation Memories"
          },
          "photoDescription": {
            "type": "string",
            "example": "A beautiful sunset at the beach"
          },
          "videoTitle": {
            "type": "string",
            "example": "Grandma’s 80th Birthday Celebration"
          },
          "isMainVideo": {
            "type": "boolean",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "mimeType",
          "fileURL",
          "type",
          "memorialId",
          "userId",
          "isMainVideo",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReadUserTributeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-tribute"
          },
          "authorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "content": {
            "type": "string",
            "example": "You will always be remembered for your kindness and love."
          },
          "type": {
            "type": "string",
            "enum": [
              "memory",
              "tribute",
              "condolence"
            ],
            "example": "tribute"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/image.jpg"
          },
          "imageId": {
            "type": "string"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "status": {
            "type": "string",
            "example": "approved",
            "description": "Moderation status"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "authorName",
          "content",
          "type",
          "imageId",
          "memorialId",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReadMemorialSayingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-saying"
          },
          "content": {
            "type": "string",
            "example": "You will always be remembered for your kindness and love."
          },
          "authorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "content",
          "memorialId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReadMemorialQRCodeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-qr-code"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "qrCodeData": {
            "type": "string",
            "example": "https://rememberme.com/memorial/john-doe"
          },
          "qrCodeImageURL": {
            "type": "string",
            "example": "https://cdn.example.com/qr-codes/qr-code.png"
          },
          "qrCodeImageId": {
            "type": "string",
            "example": "uuid-of-qr-code-image"
          },
          "description": {
            "type": "string",
            "example": "QR code for sharing memorial page"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "memorialId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReadMemorialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "description": {
            "type": "string",
            "example": "Celebrating the life of John Doe"
          },
          "favSaying": {
            "type": "string",
            "example": "John was a wonderful father and friend."
          },
          "creatorId": {
            "type": "string",
            "example": "uuid-of-creator"
          },
          "landingModeId": {
            "type": "string",
            "example": "uuid-of-landing-mode"
          },
          "landingMode": {
            "$ref": "#/components/schemas/ReadLandingModeResponseDto"
          },
          "templateId": {
            "type": "string",
            "example": "uuid-of-template"
          },
          "personName": {
            "type": "string",
            "example": "John Doe"
          },
          "personGender": {
            "type": "string",
            "example": "male",
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ]
          },
          "personBirthDate": {
            "format": "date-time",
            "type": "string",
            "example": "1950-01-01T00:00:00Z"
          },
          "personDeathDate": {
            "format": "date-time",
            "type": "string",
            "example": "2020-05-15T00:00:00Z"
          },
          "personProfilePicture": {
            "type": "string",
            "example": "https://cdn.example.com/photos/john.png"
          },
          "profilePictureId": {
            "type": "string",
            "example": "uuid-of-profile-picture"
          },
          "favQuote": {
            "type": "string",
            "example": "“Live life to the fullest.”"
          },
          "featuredPhotoId": {
            "type": "string",
            "example": "uuid-of-featured-photo"
          },
          "featuredPhotoURL": {
            "type": "string",
            "example": "https://cdn.example.com/photos/featured.jpg"
          },
          "lifeStoryText": {
            "type": "string",
            "example": "John was a wonderful father and friend..."
          },
          "lifeStoryImageId": {
            "type": "string",
            "example": "uuid-of-life-story-image"
          },
          "lifeStoryImageURL": {
            "type": "string",
            "example": "https://cdn.example.com/photos/life-story.jpg"
          },
          "publishStatus": {
            "type": "string",
            "example": "draft",
            "enum": [
              "draft",
              "standard",
              "private",
              "archived"
            ]
          },
          "eventStart": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T10:00:00Z"
          },
          "eventDuration": {
            "type": "string",
            "example": "48h"
          },
          "autoRevertToFullMode": {
            "type": "boolean",
            "example": true
          },
          "slug": {
            "type": "string",
            "example": "john-doe-memorial"
          },
          "pageURL": {
            "type": "string",
            "example": "https://rememberme.com/memorial/john-doe"
          },
          "landingModeType": {
            "type": "string",
            "example": "full-mode",
            "enum": [
              "full-mode",
              "event-mode",
              "video-only-mode"
            ],
            "description": "Memorial mode type from landing mode"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-29T09:21:45Z"
          },
          "userMedia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadUserMediaResponseDto"
            }
          },
          "userTributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadUserTributeResponseDto"
            }
          },
          "favoriteSayings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadMemorialSayingResponseDto"
            }
          },
          "qrCode": {
            "$ref": "#/components/schemas/ReadMemorialQRCodeResponseDto"
          }
        },
        "required": [
          "id",
          "creatorId",
          "landingModeId",
          "templateId",
          "autoRevertToFullMode",
          "landingModeType",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadUserFeaturedMemorialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-featured-memorial"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "isFeature": {
            "type": "boolean",
            "example": true,
            "description": "Whether the memorial is currently featured"
          },
          "memorial": {
            "$ref": "#/components/schemas/ReadMemorialResponseDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "userId",
          "memorialId",
          "isFeature",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadUserNotificationSettingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "IsInAppAllowed": {
            "type": "boolean"
          },
          "IsEmailAllowed": {
            "type": "boolean"
          },
          "IsDesktopNotificationAllowed": {
            "type": "boolean"
          },
          "IsUnreadBadgeEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "userId",
          "IsInAppAllowed",
          "IsEmailAllowed",
          "IsDesktopNotificationAllowed",
          "IsUnreadBadgeEnabled",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateUserAnswerRequestDto": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string",
            "description": "The user answer text"
          }
        },
        "required": [
          "answer"
        ]
      },
      "CreateUserAnswersResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateUserAnswerRequestDto": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string",
            "description": "The user answer text"
          }
        }
      },
      "UpdateUserAnswersResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "ReadUserAnswerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-answer"
          },
          "answer": {
            "type": "string",
            "example": "I exercise 3 times a week."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T12:34:56Z"
          }
        },
        "required": [
          "id",
          "answer",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}